+++ semana19/README.md
@@ -1,0 +1,2 @@
+# LABOOK
+- Este é um projeto básico de Backend, utilizando Typescript, MySQL, além das bibliotecas knexJs e expressJs, desenvolvido para atender aos modelos REST API.
