+++ semana19/projeto/src/index.ts
@@ -1,26 +1,12 @@
 import express, {Express} from 'express'
 import cors from 'cors'
 import { AddressInfo } from "net";
-import dotenv from "dotenv";
-import knex from 'knex'
-
-dotenv.config();
 
 const app: Express = express();
 
 app.use(express.json());
 app.use(cors());
 
-export const connection = knex({
-	client: "mysql",
-	connection: {
-    host: process.env.DB_HOST,
-    port: 3306,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-  }
-});
 
 const server = app.listen(process.env.PORT || 3003, () => {
     if (server) {
